veyron-browser: Add simple RPCs and local storage.
When clicking on a method signature that takes 0 parameters,
an RPC call is made on that service. Results are logged to the console.

To demonstrate the use of local storage, the last namespace route browsed
is stored in local memory and replaces the default index page.

Local storage uses JSON.stringify and JSON.parse in order to store any type
of value, not just strings.

Change-Id: I87bd4bb97259996859a74d03b3c003161645085d