veyron-browser: Add library tests
Unit tests for exists and local-storage are added.

The other files will not have tests added in this CL.
document-ready doesn't do anything special.
isMounttable's behavior needs to be changed, so it cannot be tested yet.
mountpoint and promises are supposed to be added to veyron.js, so
they will presumably be tested there.

Change-Id: I8e81010cf1fc5f1243d4007cc9773f8d6d0340c1