veyron-browser: Add initial logging and learning service
Include the initial logging and learning hooks for later improvement.
Learning will allow automatic shortcut suggestion and automatic gathering of
additional service details.

learner-shortcut; Does a sort of weighted sum of paths to determine importance.
learner-autoRPC; Uses perceptron learning to determine when to perform the RPC automatically.

These algorithms are mostly heuristics. Until there is a mechanism for obtaining
negative feedback, the performance is not intended to be perfect.

Tests are included for the learning library and smart-service.

Change-Id: I8870f32e769811d1c01ea92d00e6856365a22c76