veyron-browser: Make RPC form input and output
The RPC form will allow the ability to make RPCs that take inputs.
Due to VOM limitations, input and output need to be simple types.
(Input/output cannot use custom-defined data structures.)

The RPC form also has a section for displaying outputs.

The selected input method and outputs are cleared whenever we access a
different name's details page.

The form's styling is a WIP. It does not look great, but it is functional.

Change-Id: If3f42cb335217e103114333e5f48fbb299677b8b