veyron-browser: Add a section to display shortcuts.
The smart-service was updated to better rank and suggest shortcuts.

Data is gathered from the names that the user browses to.
The shortcuts shown are refreshed whenever the user browses to a name.

There isn't support yet for a variable # of shortcuts. (It's fixed at 3.)
We only have shortcuts to mounttable/intermediate nodes.
There are no shortcuts for service objects yet.

The latter would require adjusting where we log and update shortcuts.
The best event handler candidate seems to be the display-item-details event.

Change-Id: I9c3ed37e4c021e0ba60820803ca67f3f8a3a55e5