veyron-browser: Suggest RPCs to perform
Instead of automatically performing RPCs, the browser will now recommend them.
When the browser is quite confident, it will make a suggestion with timeout.
After 5 seconds, the RPC will be made unless the user cancels it.

The idea seems sound, but the algorithms backing the recommendations could be
improved. We could:
* filter the methods suggested. (i.e., only recommending ReadLabel methods)
* add a threshold to decide to act without the timer.
* change the features used for prediction
* change the algorithm from perceptron to something else

Change-Id: Ib3f2c459ce7e7868c546f961c62c85a6f7ac1cfd