veyron-browser: Plugin system for RPC output
Originally, JSON.stringify printed out all values regardless of type.
However, RPC output sometimes should be rendered differently.

In the case of monitoring, we receive an object, a histogram.
The histogram, when detected, will be rendered specially.
This is the motivation for a new plugin system for rendering RPC output

The new plugin system consists of 2 parts
* Given a value, identify the correct plugin to use
* Use the plugin to render the value

The code has been restructured to accommodate plugins for rendering output.

Change-Id: I910303d33582b563544d237c9027a539853f86bf