er: Refactoring the glue service between the UI and Veyron and adding e2e integration tests.
services/namespace:
-Use of mercuries immutable data structures instead of regular
plain old js objects. For example:
-Glob() now returns an ObservableArray instead of Array. This allows
us to easily switch to watchGlob and get live updates in the UI also
this allows us to add items to UI as data becomes available instead of
waiting for glob to finish. Glob is a streaming method and should be
treated as such.
-Data objects return from Glob() (e.g. a  namespace-item) is now
explicitly defined in item.js and again uses immutable data structures.
-These changes also allowed us to load more information about an item
asynchronously (e.g. Whether it's a server or an intermediary node, etc..)
before returning it to the UI. This would eliminate the previous flickery
behaviour where a node was added to UI and then later was changed to be
browsable or not.
-Using more flatten promise chaning than previous code.
-Tests were split into unit and integration.
-e2e intgeration tests were added for the new namespace service.

Notes:
-UI is still using the old browse-service. I have not switched it yet
to the new service.
-Running integration tests requires manually starting the daemons needed
for testing by calling run-test-services.sh. This obviously suck but I am
waiting for CL/4316 that is trying to solve this problem for Veyron.JS and
to reuse as much code/pattern as possible from that approach so this is
temporary
-services.sh was split into run-webapp-services.sh, run-test-services.sh, and run-common.sh
-Makefile was updated to reflect the unit-test and integration-test split
Same for the JS test files

TODOs to follow this CL:
-Updating UI to use the new namespace/service.js and remove browse-service.js
and related unused code.
-Align the integration test approach with Veyron.JS

Change-Id: Ied8833853236651db1a30c755c32e299a0b04c90