veyron-browser: Make Autocomplete Web Component
This CL introduces an autocomplete web component.

We need a web component with the following properties
* suggests values to the user as they type
* values may be updated asynchronously
* styled similarly to paper-input with the potential for custom css
* works with mercury

Optional properties included
* use keyboard to select suggestions
* hover over and click suggestions
* change events should fire when appropriate
* input events should fire when appropriate

The approach taken is to create a web component wrapper around
paper-input and various overlay/list Polymer components.
Here, paper-autocomplete extends paper-input.

I believe that the functionality is present and that most bugs are
taken care of. The biggest known bug is between Mercury and Polymer.
The web components use polyfill to create the shadow DOM, but their
events are still accessible from the top level when listening in
capturing mode.

That is, since mercury always uses the following form,
  X.addEventListener(ev-type, handler, true)
previously internal events to Polymer components fire in addition to
the parent components because stopPropagation() cannot be used (until
too late) on the events during capturing mode.

It would also be nice to have automatic test cases for this component.
Testing it all manually is both inconsistent and time-consuming.

Change-Id: Ic5aec635333eafc04de68218bfc095e2d7b9461a