veyron-browser: Add autocomplete implementation
With the paper-autocomplete web component, it is time to gather
and suggest values for each input.

There are 3 current inputs
* name; suggest based on the glob at that level, triggered by '/'
* glob; unclear
* methodInput: suggest based on previous entries to this input

methodInput will be modified to have suggestions based on a learner.
For now, we require an exact signature, methodName, and argName match.
The perceptron-rule will update each suggestion's rank.
When predicting, the best suggestions will be returned.

In the future, it may be nice to generalize even when the signature,
methodName, and argName don't match exactly.

Change-Id: Ie293958e45b955c6ac2c0b10fed0baf034a772b3