veyron-browser: Having Veyron-Browser build the binaries it needs (wspr, mounttabled, sampled, etc.. ) in its own bin directory by setting GOBIN before doing an install. This eliminates the need to know where veyron services binaries are installed by default.
also refactoring the shell code to adhere to Google guidelines and
removing a few sleep calls for better stability.

as part of compilation of dependent services is moved to shell as well
and "make" does not compile those dependencies anymore. That make sense
given those dependencies are meant for testing and are not part of
the apps build/compile cycle.

Change-Id: I8e3d8cd9deb8eab5b8201604c4a95a72a89d37f6