veyron-browser: Autocomplete with append mode
A delimiter attribute was added to the autocomplete element.
This allows for special replacement of the input values.

The allowWhitespace attribute adds to the delimiter, by ignoring
whitespace after the delimiter.

This augmented element is now used in the Veyron browser to suggest
children to be globbed to next upon the user pressing '/'.

Additionally, the PaperInputValueEvent was renamed to something a
little more generic.

Change-Id: I3b1318e585ae737fc1e1c6aa2c5af35fecd53fcd