veyron-browser: Always refresh methodForm state
I wasn't refreshing before, and this led to a bug where different
services with the same method name wouldn't refresh. Making things a
little weird.

Solutions were to key based on itemName+methodName or just to remove
the caching. Since display-item-details shouldn't be called much,
it's fine to just not cache anymore.

Change-Id: I0bd6bfbe52d7424b82e1ca2a936f4f4d97097699