veyron-browser: Make invocation recommendations
Use the smart-service to recommend method invocations too.
This paves the way for the starring/pinning system.

Screenshots
- https://screenshot.googleplex.com/nZ401VR0oe (2 recommendations)
- https://screenshot.googleplex.com/GtsBO097o1 (if 1 of the 2 got starred)

This change also shows how the starring system would work.
However, the stars are not maintained in local storage.
Only recommendations can be starred at the moment.
Inputted method invocations have no mechanism for getting stars.

There's no limit to the number of starred items that can be shown.
The <= 2 recommended items will not be shown if it'd result in >3 items.

Change-Id: Ic3d20e329632a2c8fbc42a59ca8160e55c57752b