veyron-browser: Making the observable array returned by our services to be immutable.
Returning a mutable observ-array as we did before was really bad since the caller
could modify stuff and considering we were caching the object, next call to the service
would have turned a modified strcuture which did not reflect the actual results.

We should make sure any observ-x we return from our services is immutable.

There is no need to make items being push to this array immutable however since
only get to get to them is by observing the array which returns immutable versions
of the items anyway.

Change-Id: I78895744d0a3d2f96c36605c969711cdb5c479af