veyron-browser: Ensuring freezing of mutable observables into immutable ones keeps the methods that can not mutate, such as map, filter, etc.
My previous change of freezing the result of glob() broke auto-complete
since it expected .map() to be present but when I was freezing the objects
I was removing evey method from it. This change ensure we keep the methods
that can not mutate the object on it. Also ensuring thowse methods themselves
return immutable observables.

Change-Id: Ie0b3859dbf7e1b14a41f2d56d4041e88e025b494