veyron-browser: Removing all but two references to the old service layer.
Deleting mountpoint and related libraries as nothing depends on them anymore.

Note the addtion of mercury.watch in loadSubNodes in the visualize component
is needed because mercury by default does not call the listener added to
an observable until the next change. mercury.watch can be used so that
listener is called immediately. See:
https://github.com/Raynos/mercury/blob/2d32318ebdff304fad2c0c2dedcddda99be81d20/docs/observ.md#what-about-dominictarrobservable-

Change-Id: I3e6bd07e956635dfdfc509c98e9537f9887dd15c