veyron-browser: Async Smart Service
The smart service has been changed to use Promises.
- implementation updated to use promises
- tests updated to take this into account
- uses of the smart service also changed

The smart service was reworked;
'record' => 'update' (which also calls save)

This CL solves the shortcut bug as well, by caching the promises of
the learners instead of caching the learners directly.

Tests were also added to exercise the Promise rejections in the smart
service.

Change-Id: Ica4c6ff04de8d5dbffca4b3abd72b990bc337de3