veyron-browser: Update uses of Signature to Map
Since Maps are now properly vom encoded/decoded as ES6 Map, I am
updating the implementation to correspond with this change.

Tiny fix for shortcutKeys. If none were in the store, we would never
be able to save any. Added an initializer.

Also added esnext option, so that we can start using for...of

Change-Id: Ie876f60daa22c3f4135f04c6f6cf17e421d64c53