veyron-browser: Stop Chrome from Refreshing on certain href
When using the pound key for href, one usually does not expect the
page to refresh. It seems likely that Chrome now always refreshes.

Since the original intent was to simply show a link-cursor when
hovering over clickable items, we now use javascript:; instead.

Note that this means we must allow scripturl in jshint.

Change-Id: Ic10fc0cca69453a6e31df0fc4fdbcc3a525402da