veyron-browser: TBR: Fixing build with incorporating new Glob API.
New changes to Glob API removed the Glob method from signature.
Therefore we no longer can look for it in signature to decide
whether something is globbale or not.
Also with recent changes technically right now every server is globbable
so our definition of globbable is whether the server in question
has any children.
We may want to consider exposing some metadata about a service on
whether that service actually implements Glob or GetChildren
interfaces in a custom way or not.

Change-Id: I38d39016924da70a18212fd52a524c1936183c7c