viz: Fix namespace root's paper-autocomplete
The paper-autocomplete Was not listening to input event properly.
It seemed to be firing too early, before it updated its internal value.
This was a bug with probably Polymer and mercury.
Mercury's capturing mode messes with Polymer.

Anyway, waiting 0ms before retrieving the value seems to be the right
solution, in this single instance.

Change-Id: I056738577d46a03875a88c1feb89ebbbce890cc7