viz: Namespace Box preserves view type
It was always switching back to grid, but now it will keep.

Instead of proceeding with the previous strategy, we will only send over
the old viewType in each route.

This way, routes/browse, routes/bookmarks, and routes/recommendations all
enforce 'keep the original view'. On first load, original view is null/undefined,
but this is okay; browse-namespace will assign 'grid'.

Change-Id: If9ce16188389b7d922c1e02e4dacec35f05cdd6f