viz: Show loading icon for the tree view
Loading icons appear when loadChildren is in-progress.
Disappear after the 'end' signal is received in loadChildren.

This seems to be the correct behavior, but we don't have much to
test it on.

It's easier to show this in person, but here is a screenshot.
https://screenshot.googleplex.com/OKgTcUingu

Use 'once' instead of 'on' to listen to getChildren's 'end' event.
Note: Tests were not changed. This tweak only affects memory leakage.

Change-Id: Ife0efbb73104f8da6ef5b2a81a9aa0f8c8b2b31d