viz: Add Browse to Parent action
Add a new action that lets the user browse up the current namespace's
parent on the item details menu.
Note: This action is slightly different from the others since it only
cares about the current namespace and not the selected item.

However, this provides the functionality of 'up 1 folder'.
Whatever is currently being displayed, we can go 'up 1 folder'.

Screenshots:
- https://screenshot.googleplex.com/JnAW0u94EE (up in the tree view)
- https://screenshot.googleplex.com/f0Sgyx7nK8 (all 3 action icons)

Change-Id: Ibe8a6aeb2c22600fa70ea200120ed8162160f299