TBR: namespace-browser: paper-autocomplete delete-child event
And also highlight seeking

Added the 'delete-item' event when the user presses Shift+Delete

Unfortunately, this required me to remove the default case in e.keyCode
since it triggers on 'shift'.
On the positive side, the CL now preserves the highlight index by
seeking its updated index after every input event.

This was already reviewed here, but there's a bug with submission.
https://vanadium-review.googlesource.com/#/c/4651/

Change-Id: I55891fcd0b98c85c2bce583e26793d1748c601e1