nsb: Ensure that namespace is set from app load
Introduces services/state/service.js as the location to
manage persisted application state.

In this CL, the namespace index is loaded at the beginning of app.js
with the state service. This ensures that regardless of the initial route,
that the namespace is set to the last one the user visited.

Raises https://github.com/vanadium/browser/issues/74
Closes https://github.com/vanadium/browser/issues/48

Change-Id: I0467ea6c20bf7ce9f4e87bc4afb533655c29ae4f