namespace-browser: Sample World  in JavaScript.
This CL adds a "Demo" link and url (#/demo) that will run a sample
world service in javascript and publish it under the user's home
directory and then navigates to that directory.

The sample services were rewritten in JavaScript.
They are currently under house2 and cottage2 to avoid conflict with
the Go services.

Closes https://github.com/vanadium/browser/issues/64

Change-Id: I05d0a7961b8624398f7757e2083417b8863d0dc0