nsb: Persist the last selected view type
The state service has been augmented to persist the most
recently used view type. It also now does the same for
the side panel width.

A minor bug was fixed where we checked falsiness instead
of 'null' ness. We now only replace the value drawn from
the store if the value is null because '' is a valid value.

Closes https://github.com/vanadium/browser/issues/74

Change-Id: Iec9a4a336a5b6a422f4f6ebd353843820ec6c933